- make separate package for parser - functions getallrestaurants and getProductsFromRestByID now get URL from m MenuParser structure - parsing delay(sec), url and format string for access to menu now getting from .env file
